Move local code linter tests to separate module
