+++ tests/unit/source_code/linters/test_files.py
@@ -5,7 +5,7 @@ import pytest
 from databricks.labs.blueprint.tui import MockPrompts
 
 from databricks.labs.ucx.source_code.base import CurrentSessionState
-from databricks.labs.ucx.source_code.graph import DependencyResolver, SourceContainer
+from databricks.labs.ucx.source_code.graph import DependencyResolver
 from databricks.labs.ucx.source_code.notebooks.loaders import NotebookResolver, NotebookLoader
 from databricks.labs.ucx.source_code.linters.files import NotebookMigrator
 from databricks.labs.ucx.source_code.python_libraries import PythonLibraryResolver
@@ -19,7 +19,7 @@ from databricks.labs.ucx.source_code.files import FileLoader, ImportFileResolver
 from databricks.labs.ucx.source_code.linters.folders import LocalCodeLinter, LocalFileMigrator
 from databricks.labs.ucx.source_code.linters.context import LinterContext
 from databricks.labs.ucx.source_code.path_lookup import PathLookup
-from tests.unit import locate_site_packages, _samples_path
+from tests.unit import locate_site_packages
 
 
 def test_notebook_migrator_ignores_unsupported_extensions() -> None:
@@ -98,51 +98,6 @@ def test_migrator_walks_directory() -> None:
     migrator.apply(path)
     languages.fixer.assert_called_with(Language.PYTHON, 'some-code')
     assert languages.fixer.call_count > 1
-
-
-@pytest.fixture()
-def local_code_linter(mock_path_lookup, migration_index):
-    notebook_loader = NotebookLoader()
-    file_loader = FileLoader()
-    folder_loader = FolderLoader(notebook_loader, file_loader)
-    allow_list = KnownList()
-    pip_resolver = PythonLibraryResolver(allow_list)
-    session_state = CurrentSessionState()
-    import_file_resolver = ImportFileResolver(file_loader, allow_list)
-    resolver = DependencyResolver(
-        pip_resolver,
-        NotebookResolver(NotebookLoader()),
-        import_file_resolver,
-        import_file_resolver,
-        mock_path_lookup,
-    )
-    return LocalCodeLinter(
-        notebook_loader,
-        file_loader,
-        folder_loader,
-        mock_path_lookup,
-        session_state,
-        resolver,
-        lambda: LinterContext(migration_index),
-    )
-
-
-def test_linter_walks_directory(mock_path_lookup, local_code_linter) -> None:
-    mock_path_lookup.append_path(Path(_samples_path(SourceContainer)))
-    path = Path(__file__).parent / "../samples" / "simulate-sys-path"
-    paths: set[Path] = set()
-    advices = list(local_code_linter.lint_path(path, paths))
-    assert len(paths) > 10
-    assert not advices
-
-
-def test_linter_lints_children_in_context(mock_path_lookup, local_code_linter) -> None:
-    mock_path_lookup.append_path(Path(_samples_path(SourceContainer)))
-    path = Path(__file__).parent.parent / "samples" / "parent-child-context"
-    paths: set[Path] = set()
-    advices = list(local_code_linter.lint_path(path, paths))
-    assert len(paths) == 3
-    assert not advices
 
 
 def test_triple_dot_import() -> None:
